requirements_classif_final_product.txt
attrs==24.2.0
Cartopy==0.24.1
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
coloredlogs==15.0.1
colour==0.1.5
contourpy==1.3.0
cycler==0.12.1
fiona==1.10.1
flatbuffers==24.3.25
fonttools==4.54.1
h5py==3.12.1
humanfriendly==10.0
imageio==2.36.0
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
onnxruntime==1.19.2
packaging==24.1
pathlib==1.0.1
pillow==10.4.0
protobuf==5.28.2
pyparsing==3.2.0
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
susi==1.4.2
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.66.5
